impl features of ckb-gen-types.
