+++ util/gen-types/src/extension/calc_hash.rs
@@ -1,10 +1,9 @@
-use crate::{
-    core::ExtraHashView,
-    generated::packed,
-    prelude::*,
-    util::hash::{blake2b_256, new_blake2b},
-    vec::Vec,
-};
+#[cfg(not(feature = "std"))]
+use crate::util::hash::{blake2b_256, new_blake2b};
+#[cfg(feature = "std")]
+use ckb_hash::{blake2b_256, new_blake2b};
+
+use crate::{core, packed, prelude::*, vec::Vec};
 
 /*
  * Calculate simple hash for packed bytes wrappers.
@@ -264,7 +263,7 @@ impl<'r> packed::BlockReader<'r> {
     /// - If there is no extension, extra hash is the same as the uncles hash.
     /// - If there is a extension, then extra hash it the hash of the combination
     /// of uncles hash and the extension hash.
-    pub fn calc_extra_hash(&self) -> crate::core::ExtraHashView {
+    pub fn calc_extra_hash(&self) -> core::ExtraHashView {
         crate::core::ExtraHashView::new(self.calc_uncles_hash(), self.calc_extension_hash())
     }
 
@@ -289,7 +288,7 @@ impl_calc_special_hash_for_entity!(Block, calc_header_hash);
 impl_calc_special_hash_for_entity!(Block, calc_proposals_hash);
 impl_calc_special_hash_for_entity!(Block, calc_uncles_hash);
 impl_calc_special_hash_for_entity!(Block, calc_extension_hash, Option<packed::Byte32>);
-impl_calc_special_hash_for_entity!(Block, calc_extra_hash, ExtraHashView);
+impl_calc_special_hash_for_entity!(Block, calc_extra_hash, core::ExtraHashView);
 impl_calc_special_hash_for_entity!(Block, calc_tx_hashes, Vec<packed::Byte32>);
 impl_calc_special_hash_for_entity!(Block, calc_tx_witness_hashes, Vec<packed::Byte32>);
