+++ util/gen-types/src/extension/tests/mod.rs
@@ -1,4 +1,10 @@
+#[cfg(feature = "std")]
 mod calc_hash;
+#[cfg(feature = "std")]
 mod capacity;
+
+#[cfg(feature = "check-data")]
 mod check_data;
+
+#[cfg(feature = "serialized-size")]
 mod serialized_size;
