+++ util/gen-types/src/util/hash.rs
@@ -6,12 +6,19 @@
 //! * personalization: ckb-default-hash
 //!
 //! [blake2b]: https://blake2.net/blake2.pdf
+
+#![allow(dead_code)]
+
 pub use blake2b_ref::{Blake2b, Blake2bBuilder};
 
+#[doc(hidden)]
+pub const BLAKE2B_KEY: &[u8] = &[];
+/// Output digest size.
+pub const BLAKE2B_LEN: usize = 32;
 /// Blake2b personalization.
 pub const CKB_HASH_PERSONALIZATION: &[u8] = b"ckb-default-hash";
-/// The hash output on empty input.
 
+/// The hash output on empty input.
 pub const BLANK_HASH: [u8; 32] = [
     68, 244, 198, 151, 68, 213, 248, 197, 93, 100, 32, 98, 148, 157, 202, 228, 155, 196, 231, 239,
     67, 211, 136, 197, 161, 47, 66, 181, 99, 61, 22, 62,
@@ -20,7 +27,6 @@ pub const BLANK_HASH: [u8; 32] = [
 /// Creates a new hasher.
 ///
 /// This can be used to hash inputs incrementally.
-
 pub fn new_blake2b() -> Blake2b {
     Blake2bBuilder::new(32)
         .personal(CKB_HASH_PERSONALIZATION)
@@ -28,7 +34,6 @@ pub fn new_blake2b() -> Blake2b {
 }
 
 /// Hashes the slice of binary and returns the digest.
-
 pub fn blake2b_256<T: AsRef<[u8]>>(s: T) -> [u8; 32] {
     if s.as_ref().is_empty() {
         return BLANK_HASH;
